[Refactor] Separate memif functions into modules
This commit refactors the Memif source code for the media proxy by
separating Memif functions into different modules according to the
transport media protocol.
